split Landing into components
