+++ backend/.env
@@ -1,4 +1,4 @@
 PORT="3000"
-NODE_ADDRESS="concordiumwalletnode.com"
-SENDER_ACCOUNT="<input concordium wallet address>"
-ACCOUNT_SIGN_KEY="<your wallet's SignKey>"
+NODE_ADDRESS="20.73.23.240"
+SENDER_ACCOUNT="4JMSRFfvCz1BhMAZTYcYFJjEAdBu1LMoiVvPD3P63TVVvQopdi"
+ACCOUNT_SIGN_KEY="405edbb8a8b6f12154b196e5b7714b26627d00bc99cb150177c29af57c0d1d37"
