+++ backend/index.js
@@ -3,7 +3,7 @@ import express from 'express';
 import cors from 'cors'
 const app = express();
 const port = process.env.PORT || 3001;
-import {createDataTransfer} from './handler.js'
+import {createDataTransfer, lookupTransaction} from './handler.js'
 app.use(express.json());
 app.use(cors());
 
@@ -21,7 +21,14 @@ app.post("/notaryService", async (req, res, next) => {
   })
 });
 
+app.get("/txLookup", async (req, res, next) => {
+  // const { txHash } = req.body
+  const lookupResult = (await lookupTransaction())
+  console.log(lookupResult)
+  res.json(lookupResult)
+});
+
 app.listen(port, () => {
   console.log(`Notarizer endpoint listening on port:${port}`);
-  console.log(`set your extension's notary endpoint to: http://127.0.0.1:${port}/notaryService`)
+  console.log(`set your extension's notary endpoint to: http://127.0.0.1:${port}`)
 });
